+++ sponge/src/main/java/com/github/games647/changeskin/sponge/tasks/SkinInvalidater.java
@@ -26,6 +26,7 @@ public class SkinInvalidater implements Runnable {
             plugin.sendMessage(invoker, "invalidate-request");
 
             SkinData skin = plugin.getCore().getMojangSkinApi().downloadSkin(ownedSkin.getUuid());
+            plugin.cacheSponge(skin);
             SkinUpdater skinUpdater = new SkinUpdater(plugin, invoker, invoker, skin, false);
             plugin.getGame().getScheduler().createTaskBuilder().execute(skinUpdater).submit(plugin);
         }